requirements.txt
numpy>=1.18.1
scipy>=1.4.1
PyYAML==6.0
numpy_stl==3.0.1
plotly==5.13.1